fix: spilt c code into header and imp file
