+++ ClashX/goClash/build.sh
@@ -1,2 +1,1 @@
-rm -f *.h *.a
 python3 build_clash_universal.py
